+++ src/native/include/mono/utils/mono-forward.h
