+++ src/native/include/mono/utils/mono-logger.h
