+++ src/native/include/mono/utils/mono-private-unstable.h
